Replace Subcomponent with Component in data:remote module.
