+++ data/remote/src/main/java/jp/kuaddo/tsuidezake/data/remote/RemoteDataComponent.kt
@@ -1,0 +1,16 @@
+package jp.kuaddo.tsuidezake.data.remote
+
+import dagger.Component
+import jp.kuaddo.tsuidezake.data.remote.internal.di.RemoteDataModule
+
+@Component(
+    modules = [RemoteDataModule::class]
+)
+interface RemoteDataComponent {
+    val tsuidezakeService: TsuidezakeService
+
+    @Component.Factory
+    interface Factory {
+        fun create(): RemoteDataComponent
+    }
+}
