--- data/remote/src/main/java/jp/kuaddo/tsuidezake/data/remote/RemoteDataModule.kt
@@ -1,13 +1,0 @@
-package jp.kuaddo.tsuidezake.data.remote
-
-import dagger.Module
-import dagger.Provides
-import jp.kuaddo.tsuidezake.data.remote.internal.di.RemoteDataSubcomponent
-
-@Module(subcomponents = [RemoteDataSubcomponent::class])
-object RemoteDataModule {
-    @Provides
-    internal fun provideTsuidezakeService(factory: RemoteDataSubcomponent.Factory): TsuidezakeService {
-        return factory.create().tsuidezakeServiceImpl()
-    }
-}
