--- data/remote/src/main/java/jp/kuaddo/tsuidezake/data/remote/internal/di/RemoteDataSubcomponent.kt
@@ -1,18 +1,0 @@
-package jp.kuaddo.tsuidezake.data.remote.internal.di
-
-import dagger.Subcomponent
-import jp.kuaddo.tsuidezake.data.remote.internal.TsuidezakeServiceImpl
-
-@RemoteDataScope
-@Subcomponent(
-    modules = [ApiModule::class]
-)
-internal interface RemoteDataSubcomponent {
-
-    fun tsuidezakeServiceImpl(): TsuidezakeServiceImpl
-
-    @Subcomponent.Factory
-    interface Factory {
-        fun create(): RemoteDataSubcomponent
-    }
-}
