+++ data/repository/src/main/java/jp/kuaddo/tsuidezake/data/repository/RepositoryComponent.kt
@@ -5,16 +5,17 @@ import dagger.BindsInstance
 import dagger.Component
 import jp.kuaddo.tsuidezake.data.local.DaggerLocalDataComponent
 import jp.kuaddo.tsuidezake.data.local.LocalDataComponent
-import jp.kuaddo.tsuidezake.data.remote.RemoteDataModule
+import jp.kuaddo.tsuidezake.data.remote.DaggerRemoteDataComponent
+import jp.kuaddo.tsuidezake.data.remote.RemoteDataComponent
 import jp.kuaddo.tsuidezake.data.repository.internal.di.RepositoryModule
 
 @Component(
     modules = [
-        RemoteDataModule::class,
         RepositoryModule::class
     ],
     dependencies = [
-        LocalDataComponent::class
+        LocalDataComponent::class,
+        RemoteDataComponent::class
     ]
 )
 interface RepositoryComponent {
@@ -25,7 +26,9 @@ interface RepositoryComponent {
         fun create(
             @BindsInstance applicationContext: Context,
             localDataComponent: LocalDataComponent =
-                DaggerLocalDataComponent.factory().create(applicationContext)
+                DaggerLocalDataComponent.factory().create(applicationContext),
+            remoteDataComponent: RemoteDataComponent =
+                DaggerRemoteDataComponent.create()
         ): RepositoryComponent
     }
 }
